Project unit tests reorganization
This patch splits MSBuildTests and ProjectTests in several test classes,
based on the functionality of the tests (build, load/save, etc).
