+++ main/tests/MonoDevelop.Core.Tests/MonoDevelop.Core.Tests.csproj
@@ -1,5 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
-<Project DefaultTargets="Build" ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+﻿<Project DefaultTargets="Build" ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <Import Project="..\..\MonoDevelop.props" />
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
@@ -65,7 +64,6 @@
     <Compile Include="MonoDevelop.Projects\MSBuildProjectServiceTests.cs" />
     <Compile Include="MonoDevelop.Projects\MSBuildProjectTests.cs" />
     <Compile Include="MonoDevelop.Projects\MSBuildSearchPathTests.cs" />
-    <Compile Include="MonoDevelop.Projects\MSBuildTests.cs" />
     <Compile Include="MonoDevelop.Projects\ProjectCapabilityTests.cs" />
     <Compile Include="MonoDevelop.Projects\ProjectReevaluationTests.cs" />
     <Compile Include="MonoDevelop.Projects\ProjectTests.cs" />
@@ -83,6 +81,15 @@
     <Compile Include="MonoDevelop.Projects\WorkspaceTests.cs" />
     <Compile Include="MonoDevelop.Projects\ConfigurationMergingTests.cs" />
     <Compile Include="MonoDevelop.Projects\TestLanguageProject.cs" />
+    <Compile Include="MonoDevelop.Projects\ProjectBuildTests.cs" />
+    <Compile Include="MonoDevelop.Projects\ProjectLoadSaveTests.cs" />
+    <Compile Include="MonoDevelop.Projects\ProjectTargetEvaluationTests.cs" />
+    <Compile Include="MonoDevelop.Projects\ExtendedProjectPropertyTests.cs" />
+    <Compile Include="MonoDevelop.Projects\ProjectWithWildcardsTests.cs" />
+    <Compile Include="MonoDevelop.Projects\DotNetCoreProjectTests.cs" />
+    <Compile Include="MonoDevelop.Projects\PortableLibraryTests.cs" />
+    <Compile Include="MonoDevelop.Projects\GenericProjectTests.cs" />
+    <Compile Include="MonoDevelop.Projects\NetStandardProjectTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="MonoDevelop.Core\" />
