+++ main/tests/MonoDevelop.Core.Tests/MonoDevelop.Projects/ProjectTests.cs
@@ -34,6 +34,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using MonoDevelop.Projects.Policies;
 using System.Xml;
+using MonoDevelop.Projects.MSBuild;
 
 namespace MonoDevelop.Projects
 {
@@ -508,107 +509,6 @@ namespace MonoDevelop.Projects
 		}
 
 		[Test]
-		public async Task LoadPortableLibrary ()
-		{
-			string solFile = Util.GetSampleProject ("portable-library", "portable-library.sln");
-
-			Solution sol = (Solution) await Services.ProjectService.ReadWorkspaceItem (Util.GetMonitor (), solFile);
-			var p = sol.FindProjectByName ("PortableLibrary");
-
-			Assert.IsInstanceOf<DotNetProject> (p);
-
-			var pl = (DotNetProject)p;
-			Assert.AreEqual (".NETPortable", pl.GetDefaultTargetFrameworkId ().Identifier);
-
-			sol.Dispose ();
-		}
-
-		[Test]
-		public async Task BuildPortableLibrary ()
-		{
-			string solFile = Util.GetSampleProject ("portable-library", "portable-library.sln");
-			Solution sol = (Solution) await Services.ProjectService.ReadWorkspaceItem (Util.GetMonitor (), solFile);
-			var res = await sol.Build (Util.GetMonitor (), "Debug");
-			Assert.IsNull (res.Errors.FirstOrDefault ()?.ToString ());
-			sol.Dispose ();
-		}
-
-		[Test]
-		public async Task PortableLibraryImplicitReferences ()
-		{
-			string solFile = Util.GetSampleProject ("portable-library", "portable-library.sln");
-			Solution sol = (Solution) await Services.ProjectService.ReadWorkspaceItem (Util.GetMonitor (), solFile);
-			var p = (DotNetProject) sol.FindProjectByName ("PortableLibrary");
-			var refs = (await p.GetReferencedAssemblies (p.Configurations [0].Selector)).Select (r => r.FilePath.FileName).ToArray ();
-			sol.Dispose ();
-		}
-
-		[Test]
-		public void CreateGenericProject ()
-		{
-			var info = new ProjectCreateInformation ();
-			info.ProjectName = "Some.Test";
-			info.ProjectBasePath = "/tmp/test";
-			var doc = new XmlDocument ();
-			var projectOptions = doc.CreateElement ("Options");
-			var p = (GenericProject) Services.ProjectService.CreateProject ("GenericProject", info, projectOptions);
-			Assert.AreEqual ("Default", p.Configurations [0].Name);
-			Assert.AreEqual (MSBuildSupport.NotSupported, p.MSBuildEngineSupport);
-			p.Dispose ();
-		}
-
-		[Test]
-		public async Task LoadGenericProject ()
-		{
-			string solFile = Util.GetSampleProject ("generic-project", "generic-project.sln");
-
-			Solution sol = (Solution) await Services.ProjectService.ReadWorkspaceItem (Util.GetMonitor (), solFile);
-			var p = sol.FindProjectByName ("GenericProject");
-
-			Assert.IsInstanceOf<GenericProject> (p);
-
-			var pl = (GenericProject)p;
-			Assert.AreEqual ("Default", pl.Configurations [0].Name);
-			sol.Dispose ();
-		}
-
-		[Test]
-		public async Task LoadGenericProjectWithImportBeforePropertyGroup ()
-		{
-			string solFile = Util.GetSampleProject ("generic-project-with-import", "generic-project.sln");
-
-			Solution sol = (Solution)await Services.ProjectService.ReadWorkspaceItem (Util.GetMonitor (), solFile);
-			var p = sol.Items[0];
-
-			Assert.IsInstanceOf<GenericProject> (p);
-
-			var pl = (GenericProject)p;
-			Assert.AreEqual ("Default", pl.Configurations [0].Name);
-		}
-
-		[Test]
-		public void SanitizeProjectNamespace ()
-		{
-			var info = new ProjectCreateInformation {
-				ProjectBasePath = "/tmp/test",
-				ProjectName = "abc.0"
-			};
-
-			var doc = new XmlDocument ();
-			var projectOptions = doc.CreateElement ("Options");
-			projectOptions.SetAttribute ("language", "C#");
-
-			DotNetProject project = (DotNetProject) Services.ProjectService.CreateProject ("C#", info, projectOptions);
-			Assert.AreEqual ("abc", project.DefaultNamespace);
-			project.Dispose ();
-
-			info.ProjectName = "a.";
-			project = (DotNetProject) Services.ProjectService.CreateProject ("C#", info, projectOptions);
-			Assert.AreEqual ("a", project.DefaultNamespace);
-			project.Dispose ();
-		}
-
-		[Test]
 		public async Task RefreshInMemoryProjectFirstTime ()
 		{
 			// Check that the in-memory project data is used when the builder is loaded for the first time.
@@ -689,25 +589,15 @@ namespace MonoDevelop.Projects
 		}
 
 		[Test]
-		public async Task SerializedWrite ()
+		public void DefaultMSBuildSupportForNewProject ()
 		{
-			var node = new CustomItemNode<SerializedSaveTestExtension> ();
-			WorkspaceObject.RegisterCustomExtension (node);
+			var project = Services.ProjectService.CreateDotNetProject ("C#");
+			bool byDefault, require;
+			MSBuildProjectService.CheckHandlerUsesMSBuildEngine (project, out byDefault, out require);
+			Assert.IsTrue (byDefault);
+			Assert.IsFalse (require);
 
-			try {
-				string solFile = Util.GetSampleProject ("console-project", "ConsoleProject.sln");
-				Solution sol = (Solution) await Services.ProjectService.ReadWorkspaceItem (Util.GetMonitor (), solFile);
-				var p = (DotNetProject) sol.Items [0];
-
-				var op1 = p.SaveAsync (Util.GetMonitor ());
-				var op2 = p.SaveAsync (Util.GetMonitor ());
-				await op1;
-				await op2;
-				Assert.AreEqual (2, SerializedSaveTestExtension.SaveCount);
-				sol.Dispose ();
-			} finally {
-				WorkspaceObject.UnregisterCustomExtension (node);
-			}
+			project.Dispose ();
 		}
 
 		[Test]
@@ -824,141 +714,6 @@ namespace MonoDevelop.Projects
 			sol.Dispose ();
 		}
 
-		[Test()]
-		public async Task ProjectReferencingDisabledProject_SolutionBuildWorks ()
-		{
-			// If a project references another project that is disabled for the solution configuration it should
-			// not be built when building the solution as a whole.
-
-			// Build the solution. It should work.
-			string solFile = Util.GetSampleProject ("invalid-reference-resolution", "InvalidReferenceResolution.sln");
-			Solution sol = (Solution) await Services.ProjectService.ReadWorkspaceItem (Util.GetMonitor (), solFile);
-
-			var res = await sol.Build (Util.GetMonitor (), "Debug");
-			Assert.AreEqual (0, res.ErrorCount);
-
-			sol.Dispose ();
-		}
-
-		[Test ()]
-		public async Task ProjectReferencingConditionalReferences ()
-		{
-			string solFile = Util.GetSampleProject ("conditional-project-reference", "conditional-project-reference.sln");
-			Solution sol = (Solution)await Services.ProjectService.ReadWorkspaceItem (Util.GetMonitor (), solFile);
-			var p = sol.Items.FirstOrDefault (pr => pr.Name == "conditional-project-reference");
-
-			Assert.AreEqual (2, p.GetReferencedItems ((SolutionConfigurationSelector)"DebugWin").Count ());
-			Assert.AreEqual (1, p.GetReferencedItems ((SolutionConfigurationSelector)"Debug").Count ());
-
-			//We have intentional compile error in windowsLib project
-			var res = await p.Build (Util.GetMonitor (), (SolutionConfigurationSelector)"DebugWin", true);
-			Assert.AreEqual (1, res.ErrorCount);
-
-			res = await p.Build (Util.GetMonitor (), (SolutionConfigurationSelector)"Debug", true);
-			Assert.AreEqual (0, res.ErrorCount);
-
-			sol.Dispose ();
-		}
-
-		[Test()]
-		public async Task ProjectReferencingDisabledProject_ProjectBuildFails ()
-		{
-			// If a project references another project that is disabled for the solution configuration, the referenced
-			// project should build when directly building the referencing project.
-
-			// Build the solution. It should work.
-			string solFile = Util.GetSampleProject ("invalid-reference-resolution", "InvalidReferenceResolution.sln");
-			Solution sol = (Solution) await Services.ProjectService.ReadWorkspaceItem (Util.GetMonitor (), solFile);
-			var p = sol.Items.FirstOrDefault (pr => pr.Name == "ReferencingProject");
-
-			var res = await p.Build (Util.GetMonitor (), (SolutionConfigurationSelector) "Debug", true);
-			Assert.AreEqual (1, res.ErrorCount);
-
-			sol.Dispose ();
-		}
-
-		[Test]
-		public async Task UserProperties ()
-		{
-			string solFile = Util.GetSampleProject ("console-project", "ConsoleProject.sln");
-			Solution sol = (Solution) await Services.ProjectService.ReadWorkspaceItem (Util.GetMonitor (), solFile);
-			var p = (DotNetProject) sol.Items [0];
-			sol.UserProperties.SetValue ("SolProp", "foo");
-			p.UserProperties.SetValue ("ProjectProp", "bar");
-			await sol.SaveUserProperties ();
-			sol.Dispose ();
-
-			sol = (Solution) await Services.ProjectService.ReadWorkspaceItem (Util.GetMonitor (), solFile);
-			p = (DotNetProject) sol.Items [0];
-
-			Assert.AreEqual ("foo", sol.UserProperties.GetValue<string> ("SolProp"));
-			Assert.AreEqual ("bar", p.UserProperties.GetValue<string> ("ProjectProp"));
-
-			sol.Dispose ();
-		}
-
-		/// <summary>
-		/// With a PCL project having no references if you add a reference, then remove, then add it
-		/// again then the saved project file will end up no references.
-		/// </summary>
-		[Test]
-		public async Task AddingRemovingAndThenAddingReferenceToPortableLibrarySavesReferenceToFile ()
-		{
-			string solFile = Util.GetSampleProject ("portable-library", "portable-library.sln");
-
-			Solution sol = (Solution) await Services.ProjectService.ReadWorkspaceItem (Util.GetMonitor (), solFile);
-			var p = sol.FindProjectByName ("PortableLibrary") as DotNetProject;
-
-			Assert.AreEqual (0, p.References.Count);
-
-			// Add System.Xml reference.
-			p.References.Add (ProjectReference.CreateAssemblyReference ("System.Xml"));
-			await p.SaveAsync (Util.GetMonitor ());
-			Assert.AreEqual (1, p.References.Count);
-
-			// Remove System.Xml reference so no references remain.
-			p.References.RemoveAt (0);
-			await p.SaveAsync (Util.GetMonitor ());
-			Assert.AreEqual (0, p.References.Count);
-
-			// Add System.Xml reference again.
-			p.References.Add (ProjectReference.CreateAssemblyReference ("System.Xml"));
-			await p.SaveAsync (Util.GetMonitor ());
-			Assert.AreEqual (1, p.References.Count);
-
-			// Ensure the references are saved to the file.
-			sol = (Solution) await Services.ProjectService.ReadWorkspaceItem (Util.GetMonitor (), solFile);
-			p = sol.FindProjectByName ("PortableLibrary") as DotNetProject;
-
-			Assert.AreEqual (1, p.References.Count);
-			Assert.AreEqual ("System.Xml", p.References[0].Include);
-
-			sol.Dispose ();
-		}
-
-		[Test]
-		public async Task AddImportThenRemoveImportAndThenAddImportAgain ()
-		{
-			string solFile = Util.GetSampleProject ("console-project", "ConsoleProject.sln");
-			Solution sol = (Solution)await Services.ProjectService.ReadWorkspaceItem (Util.GetMonitor (), solFile);
-
-			var p = (DotNetProject)sol.Items [0];
-			p.AddImportIfMissing ("MyImport.targets", null);
-			await p.SaveAsync (Util.GetMonitor ());
-
-			p.RemoveImport ("MyImport.targets");
-			await p.SaveAsync (Util.GetMonitor ());
-
-			p.AddImportIfMissing ("MyImport.targets", null);
-			await p.SaveAsync (Util.GetMonitor ());
-
-			var savedXml = File.ReadAllText (p.FileName);
-
-			Assert.That (savedXml, Contains.Substring ("<Import Project=\"MyImport.targets\""));
-
-			sol.Dispose ();
-		}
-
 		[Test]
 		public async Task AddRemoveReferenceEvents ()
 		{
@@ -1036,207 +791,6 @@ namespace MonoDevelop.Projects
 		}
 
 		[Test]
-		public void LoadReferenceWithSpaces_bug43510 ()
-		{
-			var pref = ProjectReference.CreateAssemblyReference (" gtk-sharp, Version=2.12.0.0, Culture=neutral, PublicKeyToken=35e10195dab3c99f");
-			var p = (DotNetProject) Services.ProjectService.CreateProject ("C#");
-			p.References.Add (pref);
-			Assert.IsTrue (pref.IsValid);
-			p.Dispose ();
-		}
-
-		[Test]
-		public async Task FastBuildCheckWithLibrary ()
-		{
-			string solFile = Util.GetSampleProject ("fast-build-test", "FastBuildTest.sln");
-			Solution sol = (Solution)await Services.ProjectService.ReadWorkspaceItem (Util.GetMonitor (), solFile);
-			var app = (DotNetProject)sol.Items [0];
-			var lib = (DotNetProject)sol.Items [1];
-
-			var cs = new SolutionConfigurationSelector ("Debug");
-
-			Assert.IsTrue (app.FastCheckNeedsBuild (cs));
-			Assert.IsTrue (lib.FastCheckNeedsBuild (cs));
-
-			var res = await sol.Build (Util.GetMonitor (), cs);
-			Assert.IsFalse (res.HasErrors);
-			Assert.IsFalse (app.FastCheckNeedsBuild (cs));
-			Assert.IsFalse (lib.FastCheckNeedsBuild (cs));
-
-			var myClass = sol.ItemDirectory.Combine ("MyClass.cs");
-			File.WriteAllText (myClass, "public class MyClass { public const string Message = \"Bye\" ; }");
-			FileService.NotifyFileChanged (myClass);
-
-			Assert.IsTrue (app.FastCheckNeedsBuild (cs));
-			Assert.IsTrue (lib.FastCheckNeedsBuild (cs));
-
-			res = await lib.Build (Util.GetMonitor (), cs);
-			Assert.IsFalse (res.HasErrors);
-			Assert.IsFalse (lib.FastCheckNeedsBuild (cs));
-			Assert.IsTrue (app.FastCheckNeedsBuild (cs));
-
-			res = await app.Build (Util.GetMonitor (), cs);
-			Assert.IsFalse (res.HasErrors);
-			Assert.IsFalse (lib.FastCheckNeedsBuild (cs));
-			Assert.IsFalse (app.FastCheckNeedsBuild (cs));
-
-			sol.Dispose ();
-		}
-
-		[Test]
-		public async Task WriteAndReadEnvironmentVariableProperty ()
-		{
-			string solFile = Util.GetSampleProject ("console-project", "ConsoleProject.sln");
-			var sol = (Solution) await Services.ProjectService.ReadWorkspaceItem (Util.GetMonitor (), solFile);
-
-			var p = (DotNetProject) sol.Items [0];
-			var config = (ProjectConfiguration)p.Configurations ["Debug"];
-			config.EnvironmentVariables.Add ("Test1", "Test1Value");
-
-			await p.SaveAsync (Util.GetMonitor ());
-
-			sol = (Solution) await Services.ProjectService.ReadWorkspaceItem (Util.GetMonitor (), solFile);
-			p = (DotNetProject) sol.Items [0];
-			config = (ProjectConfiguration)p.Configurations ["Debug"];
-
-			var doc = new XmlDocument ();
-			doc.Load (p.FileName);
-			var propertyGroup = doc.DocumentElement.ChildNodes [1];
-			var environmentVariablesProperty = propertyGroup.ChildNodes.OfType<XmlElement> ().Last ();
-			var environmentVariablesChildProperty = environmentVariablesProperty.ChildNodes.OfType<XmlElement> ().First ();
-			var variableProperty = environmentVariablesChildProperty.ChildNodes.OfType<XmlElement> ().First ();
-
-			Assert.AreEqual ("Test1Value", config.EnvironmentVariables ["Test1"]);
-			Assert.AreEqual ("EnvironmentVariables", environmentVariablesProperty.Name);
-			Assert.IsFalse (environmentVariablesProperty.HasAttribute ("xmlns"));
-			Assert.IsFalse (environmentVariablesChildProperty.HasAttribute ("xmlns"));
-			Assert.IsFalse (variableProperty.HasAttribute ("xmlns"));
-
-			sol.Dispose ();
-		}
-
-		[Test]
-		public async Task WriteAndReadEnvironmentVariablePropertyForSdkProject ()
-		{
-			string solFile = Util.GetSampleProject ("console-project", "ConsoleProject.sln");
-			var sol = (Solution) await Services.ProjectService.ReadWorkspaceItem (Util.GetMonitor (), solFile);
-			var p = (DotNetProject) sol.Items [0];
-
-			string xml =
-				"<Project Sdk=\"Microsoft.NET.Sdk\" ToolsVersion=\"15.0\">\r\n" +
-				"  <PropertyGroup>\r\n" +
-				"    <OutputType>Exe</OutputType>\r\n" +
-				"    <TargetFramework>netcoreapp1.0</TargetFramework>\r\n" +
-				"  </PropertyGroup>\r\n" +
-				"  <PropertyGroup Condition=\" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' \">\r\n" +
-				"    <Value1>true</Value1>\r\n" +
-				"  </PropertyGroup>\r\n" +
-				"</Project>";
-			File.WriteAllText (p.FileName, xml);
-			sol = (Solution) await Services.ProjectService.ReadWorkspaceItem (Util.GetMonitor (), solFile);
-			p = (DotNetProject) sol.Items [0];
-			var config = (ProjectConfiguration)p.Configurations ["Debug"];
-			config.EnvironmentVariables.Add ("Test1", "Test1Value");
-
-			await p.SaveAsync (Util.GetMonitor ());
-
-			sol = (Solution) await Services.ProjectService.ReadWorkspaceItem (Util.GetMonitor (), solFile);
-			p = (DotNetProject) sol.Items [0];
-			config = (ProjectConfiguration)p.Configurations ["Debug"];
-
-			var doc = new XmlDocument ();
-			doc.Load (p.FileName);
-			var propertyGroup = doc.DocumentElement.ChildNodes [1];
-			var environmentVariablesProperty = (XmlElement)propertyGroup.ChildNodes [1];
-			var environmentVariablesChildProperty = environmentVariablesProperty.ChildNodes.OfType<XmlElement> ().First ();
-			var variableProperty = environmentVariablesChildProperty.ChildNodes.OfType<XmlElement> ().First ();
-
-			Assert.AreEqual ("Test1Value", config.EnvironmentVariables ["Test1"]);
-			Assert.AreEqual ("EnvironmentVariables", environmentVariablesProperty.Name);
-			Assert.IsFalse (environmentVariablesProperty.HasAttribute ("xmlns"));
-			Assert.IsFalse (environmentVariablesChildProperty.HasAttribute ("xmlns"));
-			Assert.IsFalse (variableProperty.HasAttribute ("xmlns"));
-
-			sol.Dispose ();
-		}
-
-		[Test]
-		public async Task NetStandardProjectReferenceIncludesFacades ()
-		{
-			// Test for https://bugzilla.xamarin.com/show_bug.cgi?id=55734
-
-			string solFile = Util.GetSampleProject ("netstandard-project", "NetStandardTest.sln");
-			var sol = (Solution)await Services.ProjectService.ReadWorkspaceItem (Util.GetMonitor (), solFile);
-
-			var p = (DotNetProject)sol.Items [0];
-			var asms = await p.GetReferencedAssemblies (p.Configurations [0].Selector);
-			Assert.IsTrue (asms.Any (r => r.FilePath.FileName == "System.Runtime.dll"));
-		}
-
-		[Test]
-		public async Task FastCheckNeedsBuildWithContext ()
-		{
-			string solFile = Util.GetSampleProject ("fast-build-test", "FastBuildTest.sln");
-			Solution sol = (Solution)await Services.ProjectService.ReadWorkspaceItem (Util.GetMonitor (), solFile);
-			var app = (DotNetProject)sol.Items [0];
-
-			var cs = new SolutionConfigurationSelector ("Debug");
-
-			var ctx = new TargetEvaluationContext ();
-			ctx.GlobalProperties.SetValue ("Foo", "Bar");
-
-			Assert.IsTrue (app.FastCheckNeedsBuild (cs, ctx));
-
-			ctx = new TargetEvaluationContext ();
-			ctx.GlobalProperties.SetValue ("Foo", "Bar");
-
-			var res = await sol.Build (Util.GetMonitor (), cs, ctx);
-			Assert.IsFalse (res.HasErrors);
-
-			ctx = new TargetEvaluationContext ();
-			ctx.GlobalProperties.SetValue ("Foo", "Bar");
-			Assert.IsFalse (app.FastCheckNeedsBuild (cs, ctx));
-
-			ctx = new TargetEvaluationContext ();
-			ctx.GlobalProperties.SetValue ("Foo", "Modified");
-			Assert.IsTrue (app.FastCheckNeedsBuild (cs, ctx));
-
-			sol.Dispose ();
-		}
-
-		[Test]
-		public async Task OnConfigureTargetEvaluationContext ()
-		{
-			var node = new CustomItemNode<EvalContextCreationTestExtension> ();
-			WorkspaceObject.RegisterCustomExtension (node);
-
-			try {
-				string solFile = Util.GetSampleProject ("fast-build-test", "FastBuildTest.sln");
-				Solution sol = (Solution)await Services.ProjectService.ReadWorkspaceItem (Util.GetMonitor (), solFile);
-				var app = (DotNetProject)sol.Items [0];
-
-				var cs = new SolutionConfigurationSelector ("Debug");
-
-				Assert.IsTrue (app.FastCheckNeedsBuild (cs));
-
-				var res = await sol.Build (Util.GetMonitor (), cs);
-				Assert.IsFalse (res.HasErrors);
-				Assert.IsFalse (app.FastCheckNeedsBuild (cs));
-
-				EvalContextCreationTestExtension.ControlValue = "Changed";
-				Assert.IsTrue (app.FastCheckNeedsBuild (cs));
-
-				res = await sol.Build (Util.GetMonitor (), cs);
-				Assert.IsFalse (res.HasErrors);
-				Assert.IsFalse (app.FastCheckNeedsBuild (cs));
-
-				sol.Dispose ();
-			} finally {
-				WorkspaceObject.UnregisterCustomExtension (node);
-			}
-		}
-
-		[Test]
 		public void GetDefaultNamespaceWhenProjectRootNamespaceContainsHyphen ()
 		{
 			var project = Services.ProjectService.CreateDotNetProject ("C#");
@@ -1244,35 +798,6 @@ namespace MonoDevelop.Projects
 			string defaultNamespace = project.GetDefaultNamespace (null);
 
 			Assert.AreEqual ("abctest", defaultNamespace);
-		}
-	}
-
-	class SerializedSaveTestExtension: SolutionItemExtension
-	{
-		static bool Running = false;
-		public static int SaveCount = 0;
-
-		internal protected override async Task OnSave (ProgressMonitor monitor)
-		{
-			if (Running)
-				Assert.Fail ("A save operation is already running");
-			Running = true;
-			await Task.Delay (500);
-			Running = false;
-			SaveCount++;
-			await base.OnSave (monitor);
-		}
-	}
-
-	class EvalContextCreationTestExtension : ProjectExtension
-	{
-		public static string ControlValue = "First";
-
-		internal protected override TargetEvaluationContext OnConfigureTargetEvaluationContext (string target, ConfigurationSelector configuration, TargetEvaluationContext context)
-		{
-			var c = base.OnConfigureTargetEvaluationContext (target, configuration, context);
-			context.GlobalProperties.SetValue ("Foo", ControlValue);
-			return c;
 		}
 	}
 }
