+++ main/tests/MonoDevelop.Core.Tests/MonoDevelop.Projects/SolutionTests.cs
@@ -442,98 +442,7 @@ namespace MonoDevelop.Projects
 
 			sol.Dispose ();
 		}
-		
-		[Test()]
-		public async Task BuildingAndCleaning ()
-		{
-			string solFile = Util.GetSampleProject ("console-with-libs", "console-with-libs.sln");
-			
-			Solution sol = (Solution) await Services.ProjectService.ReadWorkspaceItem (Util.GetMonitor (), solFile);
-			DotNetProject p = (DotNetProject) sol.FindProjectByName ("console-with-libs");
-			DotNetProject lib1 = (DotNetProject) sol.FindProjectByName ("library1");
-			DotNetProject lib2 = (DotNetProject) sol.FindProjectByName ("library2");
-			
-			SolutionFolder folder = new SolutionFolder ();
-			folder.Name = "subfolder";
-			sol.RootFolder.Items.Add (folder);
-			sol.RootFolder.Items.Remove (lib2);
-			folder.Items.Add (lib2);
-			
-			Workspace ws = new Workspace ();
-			ws.FileName = Path.Combine (sol.BaseDirectory, "workspace");
-			ws.Items.Add (sol);
-			await ws.SaveAsync (Util.GetMonitor ());
-			
-			// Build the project and the references
-			
-			BuildResult res = await ws.Build (Util.GetMonitor (), ConfigurationSelector.Default);
-			Assert.AreEqual (0, res.ErrorCount);
-			Assert.AreEqual (0, res.WarningCount);
-			Assert.AreEqual (3, res.BuildCount);
-			
-			Assert.IsTrue (File.Exists (Util.Combine (p.BaseDirectory, "bin", "Debug", "console-with-libs.exe")));
-			Assert.IsTrue (File.Exists (Util.Combine (p.BaseDirectory, "bin", "Debug", GetMdb (p, "console-with-libs.exe"))));
-			Assert.IsTrue (File.Exists (Util.Combine (lib1.BaseDirectory, "bin", "Debug", "library1.dll")));
-			Assert.IsTrue (File.Exists (Util.Combine (lib1.BaseDirectory, "bin", "Debug", GetMdb (lib1, "library1.dll"))));
-			Assert.IsTrue (File.Exists (Util.Combine (lib2.BaseDirectory, "bin", "Debug", "library2.dll")));
-			Assert.IsTrue (File.Exists (Util.Combine (lib2.BaseDirectory, "bin", "Debug", GetMdb (lib2, "library2.dll"))));
-			
-			// Clean the workspace
-			
-			await ws.Clean (Util.GetMonitor (), ConfigurationSelector.Default);
-			Assert.IsFalse (File.Exists (Util.Combine (p.BaseDirectory, "bin", "Debug", "console-with-libs.exe")));
-			Assert.IsFalse (File.Exists (Util.Combine (p.BaseDirectory, "bin", "Debug", GetMdb (p, "console-with-libs.exe"))));
-			Assert.IsFalse (File.Exists (Util.Combine (lib1.BaseDirectory, "bin", "Debug", "library1.dll")));
-			Assert.IsFalse (File.Exists (Util.Combine (lib1.BaseDirectory, "bin", "Debug", GetMdb (lib1, "library1.dll"))));
-			Assert.IsFalse (File.Exists (Util.Combine (lib2.BaseDirectory, "bin", "Debug", "library2.dll")));
-			Assert.IsFalse (File.Exists (Util.Combine (lib2.BaseDirectory, "bin", "Debug", GetMdb (lib2, "library2.dll"))));
-			
-			// Build the solution
-			
-			res = await ws.Build (Util.GetMonitor (), ConfigurationSelector.Default);
-			Assert.AreEqual (0, res.ErrorCount);
-			Assert.AreEqual (0, res.WarningCount);
-			Assert.AreEqual (3, res.BuildCount);
-			
-			Assert.IsTrue (File.Exists (Util.Combine (p.BaseDirectory, "bin", "Debug", "console-with-libs.exe")));
-			Assert.IsTrue (File.Exists (Util.Combine (p.BaseDirectory, "bin", "Debug", GetMdb (p, "console-with-libs.exe"))));
-			Assert.IsTrue (File.Exists (Util.Combine (lib1.BaseDirectory, "bin", "Debug", "library1.dll")));
-			Assert.IsTrue (File.Exists (Util.Combine (lib1.BaseDirectory, "bin", "Debug", GetMdb (lib1, "library1.dll"))));
-			Assert.IsTrue (File.Exists (Util.Combine (lib2.BaseDirectory, "bin", "Debug", "library2.dll")));
-			Assert.IsTrue (File.Exists (Util.Combine (lib2.BaseDirectory, "bin", "Debug", GetMdb (lib2, "library2.dll"))));
-			
-			// Clean the solution
-			
-			await sol.Clean (Util.GetMonitor (), "Debug");
-			Assert.IsFalse (File.Exists (Util.Combine (p.BaseDirectory, "bin", "Debug", "console-with-libs.exe")));
-			Assert.IsFalse (File.Exists (Util.Combine (p.BaseDirectory, "bin", "Debug", GetMdb (p, "console-with-libs.exe"))));
-			Assert.IsFalse (File.Exists (Util.Combine (lib1.BaseDirectory, "bin", "Debug", "library1.dll")));
-			Assert.IsFalse (File.Exists (Util.Combine (lib1.BaseDirectory, "bin", "Debug", GetMdb (lib1, "library1.dll"))));
-			Assert.IsFalse (File.Exists (Util.Combine (lib2.BaseDirectory, "bin", "Debug", "library2.dll")));
-			Assert.IsFalse (File.Exists (Util.Combine (lib2.BaseDirectory, "bin", "Debug", GetMdb (lib2, "library2.dll"))));
-			
-			// Build the solution folder
-			
-			res = await folder.Build (Util.GetMonitor (), (SolutionConfigurationSelector) "Debug");
-			Assert.AreEqual (0, res.ErrorCount);
-			Assert.AreEqual (0, res.WarningCount);
-			Assert.AreEqual (1, res.BuildCount);
-			
-			Assert.IsFalse (File.Exists (Util.Combine (p.BaseDirectory, "bin", "Debug", "console-with-libs.exe")));
-			Assert.IsFalse (File.Exists (Util.Combine (p.BaseDirectory, "bin", "Debug", GetMdb (p, "console-with-libs.exe"))));
-			Assert.IsFalse (File.Exists (Util.Combine (lib1.BaseDirectory, "bin", "Debug", "library1.dll")));
-			Assert.IsFalse (File.Exists (Util.Combine (lib1.BaseDirectory, "bin", "Debug", GetMdb (lib1, "library1.dll"))));
-			Assert.IsTrue (File.Exists (Util.Combine (lib2.BaseDirectory, "bin", "Debug", "library2.dll")));
-			Assert.IsTrue (File.Exists (Util.Combine (lib2.BaseDirectory, "bin", "Debug", GetMdb (lib2, "library2.dll"))));
-			
-			// Clean the solution folder
-			
-			await folder.Clean (Util.GetMonitor (), (SolutionConfigurationSelector) "Debug");
-			Assert.IsFalse (File.Exists (Util.Combine (lib2.BaseDirectory, "bin", "Debug", "library2.dll")));
-			Assert.IsFalse (File.Exists (Util.Combine (lib2.BaseDirectory, "bin", "Debug", GetMdb (lib2, "library2.dll"))));
-			sol.Dispose ();
-		}
-		
+
 		[Test()]
 		public async Task FormatConversions ()
 		{
@@ -591,127 +500,6 @@ namespace MonoDevelop.Projects
 
 			sol.Dispose ();
 		}
-		
-		[Test()]
-		public async Task BuildConfigurationMappings ()
-		{
-			string solFile = Util.GetSampleProject ("test-build-configs", "test-build-configs.sln");
-			
-			Solution sol = (Solution) await Services.ProjectService.ReadWorkspaceItem (Util.GetMonitor (), solFile);
-			DotNetProject lib1 = (DotNetProject) sol.FindProjectByName ("Lib1");
-			DotNetProject lib2 = (DotNetProject) sol.FindProjectByName ("Lib2");
-			DotNetProject lib3 = (DotNetProject) sol.FindProjectByName ("Lib3");
-			DotNetProject lib4 = (DotNetProject) sol.FindProjectByName ("Lib4");
-			
-			// Check that the output file names are correct
-			
-			Assert.AreEqual (GetConfigFolderName (lib1, "Debug"), "Debug");
-			Assert.AreEqual (GetConfigFolderName (lib1, "Release"), "Release");
-			Assert.AreEqual (GetConfigFolderName (lib2, "Debug"), "Release");
-			Assert.AreEqual (GetConfigFolderName (lib2, "Release"), "Debug");
-			Assert.AreEqual (GetConfigFolderName (lib3, "Debug"), "DebugExtra");
-			Assert.AreEqual (GetConfigFolderName (lib3, "Release"), "ReleaseExtra");
-
-			// Check that building the solution builds the correct project configurations
-			
-			await CheckSolutionBuildClean (sol, "Debug");
-			await CheckSolutionBuildClean (sol, "Release");
-			
-			// Check that building a project builds the correct referenced project configurations
-			
-			await CheckProjectReferencesBuildClean (sol, "Debug");
-			await CheckProjectReferencesBuildClean (sol, "Release");
-			
-			// Single project build and clean
-			
-			await CheckProjectBuildClean (lib2, "Debug");
-			await CheckProjectBuildClean (lib2, "Release");
-			await CheckProjectBuildClean (lib3, "Debug");
-			await CheckProjectBuildClean (lib3, "Release");
-			await CheckProjectBuildClean (lib4, "Debug");
-			await CheckProjectBuildClean (lib4, "Release");
-
-			sol.Dispose ();
-		}
-		
-		async Task CheckSolutionBuildClean (Solution sol, string configuration)
-		{
-			SolutionConfigurationSelector config = (SolutionConfigurationSelector) configuration;
-			string tag = "CheckSolutionBuildClean config=" + configuration;
-			DotNetProject lib1 = (DotNetProject) sol.FindProjectByName ("Lib1");
-			DotNetProject lib2 = (DotNetProject) sol.FindProjectByName ("Lib2");
-			DotNetProject lib3 = (DotNetProject) sol.FindProjectByName ("Lib3");
-			DotNetProject lib4 = (DotNetProject) sol.FindProjectByName ("Lib4");
-			
-			Assert.IsFalse (File.Exists (lib1.GetOutputFileName (config)), tag);
-			Assert.IsFalse (File.Exists (lib2.GetOutputFileName (config)), tag);
-			Assert.IsFalse (File.Exists (lib3.GetOutputFileName (config)), tag);
-			Assert.IsFalse (File.Exists (lib4.GetOutputFileName (config)), tag);
-			
-			BuildResult res = await sol.Build (Util.GetMonitor (), config);
-			Assert.AreEqual (0, res.WarningCount, tag);
-			Assert.AreEqual (0, res.ErrorCount, tag);
-			
-			Assert.IsTrue (File.Exists (lib1.GetOutputFileName (config)), tag);
-			Assert.IsTrue (File.Exists (lib2.GetOutputFileName (config)), tag);
-			Assert.IsTrue (File.Exists (lib3.GetOutputFileName (config)), tag);
-			Assert.IsTrue (File.Exists (lib4.GetOutputFileName (config)), tag);
-			
-			await sol.Clean (Util.GetMonitor (), config);
-			
-			Assert.IsFalse (File.Exists (lib1.GetOutputFileName (config)), tag);
-			Assert.IsFalse (File.Exists (lib2.GetOutputFileName (config)), tag);
-			Assert.IsFalse (File.Exists (lib3.GetOutputFileName (config)), tag);
-			Assert.IsFalse (File.Exists (lib4.GetOutputFileName (config)), tag);
-		}
-		
-		async Task CheckProjectReferencesBuildClean (Solution sol, string configuration)
-		{
-			SolutionConfigurationSelector config = (SolutionConfigurationSelector) configuration;
-			string tag = "CheckProjectReferencesBuildClean config=" + configuration;
-			DotNetProject lib1 = (DotNetProject) sol.FindProjectByName ("Lib1");
-			DotNetProject lib2 = (DotNetProject) sol.FindProjectByName ("Lib2");
-			DotNetProject lib3 = (DotNetProject) sol.FindProjectByName ("Lib3");
-			DotNetProject lib4 = (DotNetProject) sol.FindProjectByName ("Lib4");
-			
-			Assert.IsFalse (File.Exists (lib1.GetOutputFileName (config)), tag);
-			Assert.IsFalse (File.Exists (lib2.GetOutputFileName (config)), tag);
-			Assert.IsFalse (File.Exists (lib3.GetOutputFileName (config)), tag);
-			Assert.IsFalse (File.Exists (lib4.GetOutputFileName (config)), tag);
-			
-			BuildResult res = await lib1.Build (Util.GetMonitor (), config, true);
-			Assert.AreEqual (0, res.WarningCount, tag);
-			Assert.AreEqual (0, res.ErrorCount, tag + " " + res.CompilerOutput);
-			
-			Assert.IsTrue (File.Exists (lib1.GetOutputFileName (config)), tag);
-			Assert.IsTrue (File.Exists (lib2.GetOutputFileName (config)), tag);
-			Assert.IsTrue (File.Exists (lib3.GetOutputFileName (config)), tag);
-			Assert.IsTrue (File.Exists (lib4.GetOutputFileName (config)), tag);
-			
-			await sol.Clean (Util.GetMonitor (), config);
-		}
-		
-		async Task CheckProjectBuildClean (DotNetProject lib, string configuration)
-		{
-			SolutionConfigurationSelector config = (SolutionConfigurationSelector) configuration;
-			string tag = "CheckProjectBuildClean lib=" + lib.Name + " config=" + configuration;
-			
-			Assert.IsFalse (File.Exists (lib.GetOutputFileName (config)), tag);
-			
-			BuildResult res = await lib.Build (Util.GetMonitor (), config, false);
-			Assert.AreEqual (0, res.WarningCount, tag);
-			Assert.AreEqual (0, res.ErrorCount, tag);
-			
-			Assert.IsTrue (File.Exists (lib.GetOutputFileName (config)), tag);
-			
-			await lib.Clean (Util.GetMonitor (), config);
-			Assert.IsFalse (File.Exists (lib.GetOutputFileName (config)), tag);
-		}
-		
-		string GetConfigFolderName (DotNetProject lib, string conf)
-		{
-			return Path.GetFileName (Path.GetDirectoryName (lib.GetOutputFileName ((SolutionConfigurationSelector)conf)));
-		}
 
 		[Test]
 		public async Task LoadKnownUnsupportedProjects ()
@@ -741,36 +529,6 @@ namespace MonoDevelop.Projects
 
 			Assert.AreEqual (Util.GetXmlFileInfoset (p.FileName + ".saved"), Util.GetXmlFileInfoset (p.FileName));
 			Assert.AreEqual (solText, File.ReadAllLines (solFile));
-
-			sol.Dispose ();
-		}
-
-		[Test]
-		public async Task BuildSolutionWithUnsupportedProjects ()
-		{
-			string solFile = Util.GetSampleProject ("unsupported-project", "console-with-libs.sln");
-
-			Solution sol = (Solution) await Services.ProjectService.ReadWorkspaceItem (Util.GetMonitor (), solFile);
-			var res = await sol.Build (Util.GetMonitor (), "Debug");
-
-			// The solution has a console app that references an unsupported library. The build of the solution should fail.
-			Assert.AreEqual (1, res.ErrorCount);
-
-			var app = sol.GetAllItems<DotNetProject> ().FirstOrDefault (it => it.FileName.FileName == "console-with-libs.csproj");
-
-			// The console app references an unsupported library. The build of the project should fail.
-			res = await app.Build (Util.GetMonitor (), ConfigurationSelector.Default, true);
-			Assert.IsTrue (res.ErrorCount == 1);
-
-			// A solution build should succeed if it has unbuildable projects but those projects are not referenced by buildable projects
-			app.References.Clear ();
-			await sol.SaveAsync (Util.GetMonitor ());
-			res = await sol.Build (Util.GetMonitor (), "Debug");
-			Assert.IsTrue (res.ErrorCount == 0);
-
-			// Regular project not referencing anything else. Should build.
-			res = await app.Build (Util.GetMonitor (), ConfigurationSelector.Default, true);
-			Assert.IsTrue (res.ErrorCount == 0);
 
 			sol.Dispose ();
 		}
